Refactor listener functions
